+++ src/services.rs
@@ -176,6 +176,10 @@ async fn list_services(
         "minio".to_string(),
         get_minio_services(!all_namespaces, redact_credentials).await?,
     );
+    output.insert(
+        "opensearch-dashboards".to_string(),
+        get_opensearch_dashboards_services(!all_namespaces, redact_credentials).await?,
+    );
 
     match output_type {
         OutputType::Text => {
@@ -188,7 +192,7 @@ async fn list_services(
                     Cell::new("Name"),
                     Cell::new("Namespace"),
                     Cell::new("Endpoints"),
-                    Cell::new("Extra infos"),
+                    Cell::new("Info"),
                 ]);
 
             let max_endpoint_name_length = output
@@ -216,7 +220,7 @@ async fn list_services(
                         .join("\n");
 
                     table.add_row(vec![
-                        Cell::new(&product_name),
+                        Cell::new(product_name.as_str()),
                         Cell::new(installed_product.name),
                         Cell::new(installed_product.namespace.unwrap_or_default()),
                         Cell::new(endpoints),
@@ -550,4 +554,75 @@ pub async fn get_minio_service(
         endpoints,
         extra_infos,
     })
+}
+
+async fn get_opensearch_dashboards_services(
+    namespaced: bool,
+    redact_credentials: bool,
+) -> Result<Vec<InstalledProduct>, Box<dyn Error>> {
+    let client = get_client().await?;
+    let list_params = ListParams::default().labels("app.kubernetes.io/name=opensearch-dashboards");
+
+    let mut result = Vec::new();
+
+    let service_api: Api<Service> = if namespaced {
+        Api::namespaced(client.to_owned(), NAMESPACE.lock()?.as_str())
+    } else {
+        Api::all(client.clone())
+    };
+
+    let services = service_api.list(&list_params).await?;
+    for service in services {
+        let namespace = service.namespace().unwrap();
+        let endpoints =
+            get_service_endpoint_urls(&service, &service.name_unchecked(), client.to_owned())
+                .await?;
+
+        let annotations = service.annotations();
+
+        let mut extra_infos = vec!["Third party service".to_string()];
+
+        if let Some(http_endpoint) = endpoints.get("http") {
+            if let Some(logs_endpoint) = annotations.get("stackable.tech/logging-view-logs") {
+                extra_infos.push(format!("Logs view: {http_endpoint}{logs_endpoint}"));
+            }
+        }
+
+        if let Some(credentials_secret) =
+            annotations.get("stackable.tech/logging-credentials-secret")
+        {
+            let secret_api: Api<Secret> = Api::namespaced(client.clone(), &namespace);
+            if let Ok(credentials) = secret_api.get(credentials_secret).await {
+                if let Some(username) = credentials
+                    .data
+                    .as_ref()
+                    .and_then(|data| data.get("username"))
+                    .and_then(|value| String::from_utf8(value.0.to_owned()).ok())
+                {
+                    let mut credentials_info = format!("Username: {username}");
+                    if redact_credentials {
+                        credentials_info.push_str(&format!(", password: {REDACTED_PASSWORD}"));
+                    } else if let Some(password) = credentials
+                        .data
+                        .as_ref()
+                        .and_then(|data| data.get("password"))
+                        .and_then(|value| String::from_utf8(value.0.to_owned()).ok())
+                    {
+                        credentials_info.push_str(&format!(", password: {password}"));
+                    }
+                    extra_infos.push(credentials_info);
+                }
+            }
+        }
+
+        let installed_product = InstalledProduct {
+            name: service.name_unchecked(),
+            namespace: service.namespace(),
+            endpoints,
+            extra_infos,
+        };
+        result.push(installed_product);
+    }
+
+    Ok(result)
 }